services/orchest-api/app/requirements.txt
aiodocker==0.18.3
aiohttp==3.6.2
amqp==2.6.0
celery==4.4.2
docker>=4.3.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
requests==2.22.0
Werkzeug==0.16.0
python-engineio==3.13.2
python-socketio[client]==4.6.0
-e ../../../lib/python/orchest-internals
